pyproject.toml
[project]
name = "catalog"
version = "0.1.0"
requires-python = ">=3.13"
dependencies = [
    "django>=5.2",
    "django-cors-headers>=4.7.0",
    "django-ninja>=1.4.1",
    "pillow>=11.2.1",
    "psycopg[binary]>=3.2.7",
]

[dependency-groups]
dev = [
    "python-dotenv>=1.1.0",
]
